requirements.txt
streamlit==1.50.0
plotly==6.7.0
pandas==2.2.3
requests==2.32.3
python-dotenv==1.0.1
